% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinySoSV_prediction.R
\name{ShinySoSV_prediction}
\alias{ShinySoSV_prediction}
\title{ShinySoSV prediction}
\usage{
ShinySoSV_prediction(Candidate_callers, newdata, performance, callset)
}
\arguments{
\item{Candidate_callers}{names of callers}

\item{newdata}{data frame of variables with which to predict}

\item{performance}{should be one of "sensitivity", "precision" or "F1 score"}

\item{callset}{individual caller or pairwise union or intersection}
}
\value{
data frame of predicted performance
}
\description{
Performance prediction of SV calling
}
