% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_bed.R
\name{vcf_to_bed}
\alias{vcf_to_bed}
\title{Convert VCF format to bed format}
\usage{
vcf_to_bed(vcf_file)
}
\arguments{
\item{vcf_file}{names of vcf file}
}
\value{
data frame
}
\description{
This function read
}
